add UUID value to the body of the event emission
This isn't surfaced within the Datadog UI, so I'll be adding the event UUID to the output body. If cronner is ever extended to save the output of failed commands to disk, the UUID could be used as the filename. By exposing it within the event body, you can then find the specific logfile.