add option for persisting command output on failure
In some situations (big-ass stack traces), only having a Dogstatsd event emission isn't sufficient. The amount of data available is larger than we can ship as an event.

This adds the ability to persist failures to disk for later inspection. With the UUID enhancement in #14, you can easily tie an event back to a persisted file.