enhance cronner memory utilization; change saveOutput func signature
In situations where cronner didn't care about the command output, it still saved all of the command output to a buffer. This, of course, meant the output of the command always lived in memory even if we had no plans on saving/consuming it.

Using the different command line flags, we determine whether or not we care about the command output. If we do not, we set `Stdout` and `Stderr` on the command struct to nil. This effectively writes the data to `os.DevNull`.

I have no metrics on memory saving as it's absolutely dependent on the command being ran.

Also, `saveOutput()` was changed to `writeOutput()`.