prep for open sourcing cronner; add statsd namespace flag
Prior to this patch, `cronner` would set the `pagerduty` namespace on all statsd emissions. The reason for doing this was so end users within PagerDuty wouldn't need to remember to set the proper namespace. The use of the namespace was to avoid any naming collisions with metrics emitted from dd-agent checks.

With the new Chef LWRP around `cronner`, the namespace setting can be done without the humans needing to remember it. So that concern has been alleviated.

In addition, subsequent enhancements by Datadog to their service added some intelligent parsing of statsd metric names. With this change, the inadequacies of using `pagerduty.cron` become apparent as the metrics are grouped under the `pagerduty` service (as they call it). By changing this to `cronner`, it will show up as its own application. This may come in handy at some point.

There are a few other small fixes too:

* fix a spacing issue on one of the UUID lines in the event emission
* remove extra new line from the end of the help output
* improve help output by properly defining the positional arguments in the arg struct
* update documentation to reflect using positional arguments