code restructure/refactor in preparation for open source release
The code had some room for improvement. So, since I plan on trying to get this open sourced I wanted to clean it up a little bit.

This moves the command running and logging in to a single function within their own file (`runner.go`).

This also renames the cronner argument struct type to `binArgs` for sanity purposes.