refactor command running code
The command running code, initially, only handled the running of the command and figuring out run times, return code, etc. However, with the addition of the warning Datadog event, it made sense to unify these under a single function.

So, `runCommand()` has been renamed to `handleCommand()` and it takes a struct now, instead of an ever-expanding list of arguments.

As part of this, I moved the start/end event emission code in there as well as the save-output-to-disk code.

This change allows us to test the full event submission code path, so tests were added for that too.