add a group flag to cronner for further event correlation
This adds a flag to emit an additional tag on DogStatsd event emissions. This tag is `cronner_group` and it can be added by using the `-G/--event-group` flag.

If you have different cronner jobs that alert the same team, you could utilize this group for multihost Datadog event monitors.