release cronner as open source under BSD-3 Clause license
This adds a BSD 3-Clause license file, as well as updates the license header on all source files to indicate the source code is copyrighted to PagerDuty, Inc. and its usage is goverened by the BSD 3-Clause license. In addition, a contributors list has been added to the `README.md` file.

In addition, the Development instructions in the `README.md` file have been modified to remove the Godep/gpm suggestion as none of the build steps actually use `gpm` or `Godep` to ensure the right versions of the dependencies are being used. No use suggesting it if we don't actually use it.