alter argument parsing function to allow unit testing; uprev to v0.2.1
The `*binArgs.parse()` function has been modified to return a string and an error. Previously, the function would just print and exit on its own. However, by changing this behavior we can now properly test the funciton to ensure argument parsing happens as expected.

If argument parsing succeeds, and we should continue with normal operation, the string and error values will be their defaults (`""`, and `nil` respectively). If the string value is non-empty, we should print the value and exit status 0. If the error value is present, we should log the error and exit with status 1.