vendor dependencies using glide; do not use 'go get .' in tests
This change adjusts the repo to use `glide` for dependency management. In
addition to that, the dependencies themselves are vendored.

To accomplish this change, the `.travis.yml` file was changed to not use
`go get -u .` when running tests in TravisCI as the dependencies should already
be in the `vendor/` directory.