use a better method for trimming a network read buffer
`cronner` was written when I was a bit more naïve when it came to Go. Back then
I didn't consider creating a subslice of the buffer, when reading from a UDP
stream, to trim off the nil bytes, and instead used `bytes.Trim()`.

This change removes the call to `bytes.Trim()` and instead trims the byte slice
down by creating a subslice (`buffer[:n]`).

Signed-off-by: Tim Heckman <t@heckman.io>