time keeping: switch to a monotonic time source
One of `cronner`'s core features is the ability to calculate how long a command
took to run on the system. Up until now, the only time source we had available
was the Wall-Clock Time of the system where `cronner` was running.
Unfortunately, there was always the known limitation that this was subject to
the value being wrong if NTPd were to adjust the time, or if a leap second were
to occur.

CloudFlare recently had issues with a leap second causing problems in their
software stack, and in [their blog post](https://blog.cloudflare.com/how-and-why-the-leap-second-affected-cloudflare-dns/)
brought some attention to the lack of a monotonic time source in Go. There has
been an issue open in Go for awhile to add a monotonic time source to the API,
for time keeping tasks like this:

* https://github.com/golang/go/issues/12914

In this bug report above, the `github.com/aristanetworks/goarista/monotime`
package is called out as being able to provide access to the monotonic time
source.

So in short, switch our runner time keeping to use the monotonic time source
from the package above.

Signed-off-by: Tim Heckman <t@heckman.io>