add support for linking jobs running within cronner
This change makes an alteration to `cronner` which injects environment variables
to let subprocesses know they are running inside of `cronner`.

In addition to the change above, a new flag (`-P/--use-parent`) was added to
instruct `cronner` to slurp the environment variables set by the parent cronner
and emit the relevant tags with the events and metrics.

`cronner` now sets the following environment variables before invoking a
command:

* `CRONNER_PARENT_UUID`: this is the UUID of the parent `cronner` job; use this
  environment variable to determine if cronner is invoking you
* `CRONNER_PARENT_EVENT_GROUP`: this is the event group tag `cronner` is using
* `CRONNER_PARENT_GROUP`: this is the group tag `cronner` is emitting metrics
  with
* `CRONNER_PARENT_NAMESPACE`: the namespace of the parent invocation
* `CRONNER_PARENT_LABEL`: the label of the parent invocation

For events `cronner` will emit the following parent tags:

* `cronner_parent_uuid` and `cronner_parent_event_group`

For metrics `cronner` will emit the following parent tags:

* `cronner_parent_group`
* `cronner_parent_namespace`
* `cronner_parent_label`

Signed-off-by: Tim Heckman <t@heckman.io>