add Project History to README.md and clean up some sections
This change adds a Project History section to the README file now that we are
detached from the PagerDuty repo. This still links back to the original
PagerDuty repository for historical purposes.

In addition to that change, some of the phrasing in the README has been altered
to be clearer.

Signed-off-by: Tim Heckman <t@heckman.io>