use random port for listeners in tests to make them less flaky
This changes our test code to use random ports when binding a new UDP listener.
Previously we would use 8125, but would see cases where test runners would
collide trying to use that port causing flaky tests. It happened very
infrequently, but was annoying whenever it did come up.

This changes the code to use a different port for each test run, which should
avoid any weird issues with two tests trying to bind at the same time. This also
removes some janky sleep statement I added in 64906e84f0 when trying to fix this
issue the first time around back in August of 2015 (cronner v0.1.5). The sleep
statement drastically reduced how frequently we saw this issue (to the point I
thought it was no longer an issue).

Unlike the above commit, I believe this change should permanently address the
flaky tests. If not I'll be back to fix them in 2019! ;)

Signed-off-by: Tim Heckman <t@heckman.io>