Update cronner to use go1.9+ monotonic time source; prep for v1.0.0
This change updates `cronner` to switch time keeping back to the `time.Time`
type, now that go 1.9+ includes a transparent monotonic time source. There was
also a compile-time check added to hopefully prevent people from compiling
`cronner` against older versions of the Go runtime.

This change also removes the vendored dependency that was providing access to
the monotonic time source.

Finally, the version string is bumped to v1.0.0 to prepare for the 1.0 release!