Refactor the Go 1.9+ check to be more simple
The initial implementation of the Go 1.9 check duplicated the bare minimum of
`runner.go` needed to compile without any build errors. Specifically, a dummy
version of the handleCommand function had to be implemented otherwise the build
failure message was confusing.

This second iteration moves the constant definition out of `runner.go` and in to
to its own file (`go19.go`). Doing this avoided needing the `runner_nobuild.go`
file that was used. The constant check is then moved to the top of `runner.go`
with an indicator that the file has functionality requiring Go 1.9+. Doing this
check here means we can remove the build tag from `runner.go`, and rely on the
one in `go19.go`.